+++ source/app/service-providers/assets/menu.darwin.ts
@@ -151,7 +151,7 @@ export default function getMenu (
           label: trans('menu.open'),
           accelerator: 'Cmd+O',
           click: function (menuitem, focusedWindow) {
-            global.application.runCommand('open-root-file')
+            global.application.runCommand('root-open-files')
               .catch(e => global.log.error(String(e.message), e))
           }
         },
@@ -160,7 +160,7 @@ export default function getMenu (
           label: trans('menu.open_workspace'),
           accelerator: 'Cmd+Shift+O',
           click: function (menuitem, focusedWindow) {
-            global.application.runCommand('open-workspace')
+            global.application.runCommand('root-open-workspaces')
               .catch(e => global.log.error(String(e.message), e))
           }
         },